Refactor protoization functions
